interface for dao and service
