+++ batch7crud-roll5/src/main/java/com/lftechnology/batch7crud/services/EmployeeService.java
@@ -1,35 +1,21 @@
 package com.lftechnology.batch7crud.services;
 
 import java.util.List;
-import com.lftechnology.batch7crud.dao.EmployeeDAO;
+
 import com.lftechnology.batch7crud.exception.DataException;
 import com.lftechnology.batch7crud.model.Employee;
 
-public class EmployeeService {
-    EmployeeDAO employeeDAO = new EmployeeDAO();
-
-    public List<Employee> fetch(int limit, int offSet) throws DataException {
-        return employeeDAO.fetch(limit, offSet);
-    }
-
-    public Employee fetchById(int id) throws DataException {
-        return employeeDAO.fetchById(id);
-    }
-
-    public void create(Employee employee) throws DataException {
-        employeeDAO.create(employee);
-    }
-
-    public void edit(Employee employee) throws DataException {
-        employeeDAO.edit(employee);
-    }
-
-    public void deleteById(int id) throws DataException {
-        employeeDAO.deleteById(id);
-    }
-
-    public int count() throws DataException {
-        return employeeDAO.count();
-    }
+public interface EmployeeService {
+    List<Employee> fetch(int limit, int offSet) throws DataException;
+    
+    Employee fetchById(int id) throws DataException;
+    
+    void create(Employee employee) throws DataException;
+    
+    void edit(Employee employee) throws DataException;
+    
+    void deleteById(int id) throws DataException;
+    
+    int count() throws DataException;
 
 }
